DHTVirtualNode no longer has a _successor variable. Instead, the front() of the successor list is used. The code that was making sure the two were always updated simultaneously has been removed ( update_successor_list_head() ).
SuccList is made a derived class of std::list<Location> instead of
containing such a ist (_succs).

DHTVirtualNode::stabilize() throws if the successor is not set yet.

Get rid of unused _last_check_time data member in Location

Get rid of throw (dht_exception) in SuccList because it created
confusion during refactoring.

SuccList refactoring.
SuccList isStable is true if at least half of the maximum number of successors have
answered to ping during the last stabilization step

Ensure that tests return RUN_ALL_TESTS() so that they exit on error if tests fail.

Move the TransportTest and DHTTest to a header shared with the tests that need it.